pyproject.toml
[project]
name = "hr-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "python-multipart",
    "pymupdf",
    "python-docx",
    "pypdf2",
    "sqlalchemy",
    "psycopg2-binary",
    "google-genai",
    "python-dotenv",
    "tenacity>=9.1.4",
    "websockets>=13.0",
]
